Nerfed stat boosts
it's fun when playing with stat boosting moves has a real impact on the battle, especially thanks to lesser SUPER/NOT-VERY_EFFECTIVE moves.

But these stat boosting multipliers were a little too OP, with diminishing returns.

For example, if you are boosting your own evasion, or lowering your opponents accuracy or vice versa, you pretty much only need to get to boost stage 2 in order to get maximum benefit from the boost. Like with +2 stage accuracy, you'll like never miss any attack, even super inaccurate ones.

And with ATK/DEF/SPEED, just two stages and you're like already doubling the stat. So you could go from half-as-strong as your opponent to equal in only +2 boosts which is sometimes just a single move. That's crazy!

So I nerfed all this to make it more like a +2 boost will bridge the gap between pkmn that are pretty similar in stats. Like an avg stat will get boosted to be slight-above-avg.

A +4 boost will bridge the gap between a pkmn thats avg in a stat against a pkmn thats strong in a stat. (Like a pkmn with base 60 in a stat could now compare ~ to a pkmn with base 90 in that stat)

And a +6 boost stage will make an average pokemon begin to outperform against the strongest pkmn of that stat. (like base 60 in a stat will be boosted to 140).

Just to make it feel like each increment  of stat boosting feels like closing/extending the gap between another grouping of pkmn.